PR feedback * Check if `timestamp` has a proper order of magnitude in e2e test. * Renamings. * Split unit tests into 2 files.
